utils/pdf2text/requirements.txt
pdfplumber
pymupdf
pdf2image
pytesseract
pillow
tesseract
